+++ src/tasks/DockerBuildTaskDefinitionBase.ts
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ContainerPlatform } from '../runtimes/docker';
 import { DockerLabels, TaskDefinitionBase } from './TaskDefinitionBase';
 
 export interface DockerBuildOptions {
@@ -20,17 +21,3 @@ export interface DockerBuildOptions {
 export interface DockerBuildTaskDefinitionBase extends TaskDefinitionBase {
     dockerBuild?: DockerBuildOptions;
 }
-
-/**
- * Target platform for the image build
- */
-export type ContainerPlatform = undefined | {
-    /**
-     * OS of target platform
-     */
-    os?: string;
-    /**
-     * Architecture of target platform
-     */
-    architecture?: string;
-};
